Correct unit in header + remove vestigial MW variable
Molecular weight (MW) is in units of g/mol in the species dictionary (accessed here via the State_Chm object ('SpcData(IND)%Info%MW_g')), but was referred to as kg/mol in the header and one other location in gckpp_HetRates.F90. This has been udpated and the vestigial line removed. There are are local definitions of MW used here which should be updated to use species dictionary in the future.

Signed-off-by: ts551 <ts551@york.ac.uk>